Fix the bug in the functional tests
As it turns out, this was an issue due to the way testing differs from
running a build from a project. Specifically, adding an afterEvaluate
section to the plugin and putting the dependency on lcov inside there
cleaned this problem up by ensuring the dependency was present before
reaching the execution stage of the build. This has the added bonus of
allowing us to specify the lcov version in one place (i.e. the plugin's
extension).