Auto-formatted the code Reformatted the code to use the *_VERSION variable in order to make future changes in one spot only
Added logic to create the directory of where the output file will reside. Python doesn't create it and the script fails if the directories don't exist.
Fixed the paths of there the python script is located at.

Changed location of String variables to fix the issue with the closure scope.