Remove new File("...") from service distribution plugin.
The plugin was using new File("...") when it should have been using
project.file("...") instead.  The later will create a file relative to
the directory that contains the project being configured versus the
former which will result in a file relative to the directory where
the JVM/Gradle was started.  These are often not the same directory for
sub-projects.

The FT has also be updated to check to make sure the ZIP is actually
created.