Add a mocked project
We will need to test that the task is doing the right thing but, from what I
recall and from what I'm seeing online, Gradle doesn't really allow you to
create tasks except through the DSL. Thus, we'll need to figure out how to unit
test this or, if necessary, forget about unit-testing the task itself and just
unit-test helper functions. This would, then, require that we do integration and
functional tests for this, instead.