Remove monorepo task
Based on discussion with Matt and Thomas, we're going to refactor this plugin to
have a separate task for each of the stages of a release instead of trying to
keep a lot of state inside the ReleaseTask class. Thus, remove the old artifacts
of trying to replicate the ReleaseTask for a monorepo.