Make it possible to apply the mono repo release plugin
Make it possible to apply the mono repo release plugin and have it automatically
applied when the parent plugin is applied. Put everything relating to the mono
repo inside another package in the release package so that it can be applied
correctly -- is it possible to define two plugins in one META-INF properties
file?

Add a functional test for the mono repo release plugin.