Updates to FT, unit tests, and ReleaseUtil
Update the updateReleaseVersion task functional test to actually check
the last git commit message to ensure we committed the release when we
think we do. Still needs to be fixed as it doesn't work correctly, right
now.

Update the ReleaseUtil.git function to NOT take in the project since we
shouldn't be passing that around and we can just return the command
closure anyways. This should probably be done for the rest of the
ReleaseUtil functions that require the user to pass in the project -
especially the ones where it shouldn't be required.

Add unit tests for the UpdateVersionTask class.