Clean up more of the ReleaseUtil
It looks really ugly to pass null as a function's parameter, so split
out the functionality into two functions, one that accepts a
ByteArrayOutputStream and the arguments to git and another with just the
arguments to git and a newly created ByteArrayOutputStream. Then, update
the places where we were calling the git function and passing null, to
get rid of that parameter.