Fix bug causing double version update
The VersionResolver didn't have a way of getting the current project's
version without updating it based on the release type at the same time.
This is not what should happen so split that functionality into two
functions:
- getProjectVersion() which returns the project's current version
- updateProjectVersionForRelease() which updates the project's version
  based on the provided release type

Update usages of getProjectVersion() to refer to the proper function.

Minor: Use the TestingUtilities stuff to set up the VersionResolver
integration tests.