Fix the build and tests
Un-ignore the functional test for checking the updateReleaseVersion task
and fix it so that it works. We're removing the entire
sealion-java-hello-world directory after each of these tests to ensure
no state is kept over to mess up other tests. It's worth noting that I
still have a bug in the updateReleaseVersion task since it updates the
version for EACH project - I suspect this is due to the fact that the
test repo isn't really set up like a mono repo but I'll need to test
that hypothesis to be sure.

Update the MonoRepo plugin to create tasks using the correct
dependencies (i.e. at least the updateReleaseVersion task).

Allow the BumpVersionTask and CreateTagTask classes to NOT enforce the
version suffix since, in reality, the only time this should happen is on
the UpdateReleaseVersionTask.

Clean up a bit in the CreateTagTask class.

Fix Javadoc in the ReleaseUtil file.

Remove the enforceVersionSuffix parameter to getSemanticVersion since it
is a field in the class.

Fix the UpdateVersionTask unit tests to take the updates to
VersionResolver (made in a previous commit) into account.