Rename SeasideReleaseMonoRepoPlugin
Per discussion with Thomas, this is more accurate for releasing a root
project than only a mono repo. Thus, change the name to
SeasideReleaseRootProjectPlugin.

Don't apply the ReleaseRootProject plugin from the Parent plugin, for
now.

Remove dependencies section from the configureTask function since we'll
deal with the dependencies of these tasks (for now) in the pipeline
script.