Fix VersionResolver and SeasideReleaseRootProjectPluginFT
Fix the VersionResolver to point to the correct version file (i.e.
versions.gradle assumed to be at the top-level directory) for a mono
repo. Or, for a normal project, to use build.gradle as the version file.

NB: Currently, we have to make the following assumptions for this to
work correctly:
- The version file for a mono repo will always be called versions.gradle
- The version file for a mono repo will always reside in the top-level
  directory (i.e. the same directory as the mono repo's member projects)
- There will only ever be member projects one level deep in a mono repo;
  that is, a mono repo will not contain another mono repo

While limiting, for now, this allows us to find the correct version file
for the project or mono repo. That said, it is possible that adding the
version file as an extension property will fix this issue altogether.

Fix the functional tests so that they pass. We need to test the actual
functionality of the CreateTagTask, BumpVersionTask, and ReleasePushTask
in order to claim that this is actually done but, at least right now, we
know that the tasks will not fail the build.