Update the eclipse.feature plugin
Check that the feature file was copied - this helps when debugging an assertion
failure when the JAR isn't created. For example, the jar could fail to be
created because the feature.xml file isn't where it's supposed to be or for some
other reason. Now, we'll have one less reason to worry about.

Update the plugin to make the 'build' task be a default task and make 'build'
depend on the 'createJar' task. I thought this was already done but, apparently
not.

Finally, remove the default implementation of the 'copyFeatureFile' task so that
the user must configure the task to expand whatever properties they want.