Fix lots of issues.
The CI plugin no longer applies the checkstyle plugin.  That is left to the actual seaside checkstyle plugin.
The CI plugin now actually checks to see if checkstyle is configured before messing with the extensions.
The CI plugin now tolerates tasks that may not exists.  It no longer fails if the CI plugin is applied before other plugins.
The FTs for the CI plugin pass.
The compile error for the FT for the m2repo FT has been fixed.
The Jenkinsfile now sets the nexus username and password when running the FTs.