Add implementations
Add an implementation of DependencyLockPlugin based that will activate
dependency locking for each configuration and will create / configure the
resolveAndLockAllDependencies task.

Add an implementation of ResolveAndLockAllDependenciesTask that will ensure
--write-locks was passed when the task is called and, if it was, will resolve
and write lock files for each of the configurations.

Note: While I everything suggests this should be working, it's not
because (apparently) the plugin can't be found. I'll have to look into fixing
this next.