Update DependencyLockPlugin, tasks, and tests
Move the TASK_GROUP from ResolveAndLockAllDependenciesTask.java to
DependencyLockPlugin.java. Note that, while we've been putting task names in the
plugin, itself, I think it makes more sense to have a task name and it's
description be part of the task itself.

Update integration test to reference the task group constant in
DependencyLockPlugin.

Add functional a test to verify positive functionality and fix the functional
test that verifies the "--write-locks" flag is passed along with the task.

Create the actual task of type ResolveAndLockAllDependenciesTask rather than a
task of type DefaultTask.